sync user working, add counts of sB and update sec M strings

+++ app/src/main/java/edu/aku/hassannaqvi/uen_tmk/activities/SectionJActivity.java
@@ -1137,7 +1137,7 @@ public class SectionJActivity extends Activity {
             } else {
                 tj13m.setError(null);
             }
-            if (Integer.valueOf(tj13m.getText().toString()) > months) {
+            if (Long.valueOf(tj13m.getText().toString()) > months) {
                 Toast.makeText(this, "ERROR(Invalid): " + getString(R.string.month), Toast.LENGTH_SHORT).show();
                 tj13m.setError("Invalid Months!");    // Set Error on last radio button
